move some tests to train v2
